feat: separate field config from styling config
