+++ packages/frontend/src/features/sqlRunner/store/barChartSlice.ts
@@ -103,19 +103,15 @@ export const barChartConfigSlice = createSlice({
             action: PayloadAction<{ index: number; label: string }>,
         ) => {
             if (!config) return;
-            if (!config.display) {
-                config.display = {
-                    yAxis: [
-                        {
-                            label: action.payload.label,
-                        },
-                    ],
-                };
-            }
-            if (config?.display?.yAxis) {
-                config.display.yAxis[action.payload.index].label =
-                    action.payload.label;
-            }
+
+            config.display = config.display || {};
+            config.display.yAxis = config.display.yAxis || [];
+
+            const { index, label } = action.payload;
+            config.display.yAxis[index] = {
+                ...config.display.yAxis[index],
+                label,
+            };
         },
         setSeriesLabel: (
             { config },
